First steps to try to speed up full file indexing
The full file gets progressively slower as we go on,
even with the non-primary keys deleted, to the point
where it pittered out at about 2m items.

This change pulls the full-file handling out into its
own method to better control index stuff, and also
removes/replaces even the htid unique index on the htid
table to try to speed things up by ignoring any possible
key collisions (which shouldn't even happen in the full
file).

If this isn't fast enough, we're going to have to look
at maybe downloading the full file to a tab-delimited
file and doing a bulk import via a shell command.